requirements.txt
ExifRead==3.0.0
PyYAML==6.0.1
matplotlib~=3.7.2